requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
gunicorn>=21.2
openai>=1.30
azure-identity>=1.16
pydantic-settings>=2.2
pytest>=8.0
httpx>=0.27
